changed double to int64_t for third-repetitions test

+++ bitboard.hpp
@@ -38,7 +38,7 @@ private:
 	uint64_t vec1_cells[64];
 	uint64_t bitboard[32][8][8];
 	double kingSecurityArray[64][64];
-	double pawnStructureEvalute();
+	int64_t pawnStructureEvalute();
 
 	void zobristGenerator();
 	uint8_t popcount64(uint64_t value);
@@ -84,8 +84,8 @@ public:
 	bool whiteMove, passant_enable;
 
 	void evaluteAll();
-	double evalute;
-	double getEvalute();
+	int64_t evalute;
+	int64_t getEvalute();
 	void totalStaticEvalute();
 
 	uint8_t getFigure(uint8_t y, uint8_t x);
@@ -104,7 +104,7 @@ public:
 
 	bool testOfDraw();
 
-	double kingSecurity();
+	int64_t kingSecurity();
 
 	std::multiset<uint64_t> gameHash;
 };
